Move method to count records in marc file into utilities class add some tests
