--- src/main/java/org/folio/service/s3processing/MarcRawSplitterService.java
@@ -1,10 +1,0 @@
-package org.folio.service.s3processing;
-
-import io.vertx.core.Future;
-
-import java.io.InputStream;
-
-public interface MarcRawSplitterService {
-
-  Future<Integer> countRecordsInFile(InputStream inStream);
-}
